Add files via upload
Student Managment System : 
This is a c++ project. This project can be used in schools, tution, coaching classes in order to save the record of their student.
User can add a student record, search the student, update the details of a student, can display the detail of a student, can delete the record of a student.

Thank You.